Remove generic from CharEncoding
DexUtil.CharEncoding no longer uses a generic type. All methods relating to it now use BigIntegers:
- pkuObject now stores trash bytes in BigInteger arrays.

Also renamed the Nickname module to the Encoded_Nickname module.

+++ Formats/pku/pkuObject.cs
@@ -125,10 +125,10 @@ public class pkuObject : FormatObject
     public class Trash_Bytes_Class : pkuDictionaryTag
     {
         [JsonProperty("Nickname")]
-        public ushort[] Nickname { get; set; }
+        public BigInteger[] Nickname { get; set; }
 
         [JsonProperty("OT")]
-        public ushort[] OT { get; set; }
+        public BigInteger[] OT { get; set; }
     }
 
     public class Move : pkuDictionaryTag
